Change to long() to enable training reproduction for the nar model

+++ utils/nar_beam_search.py
@@ -81,7 +81,7 @@ class Beamsearch(object):
         self.next_nodes.append(new_nodes)
         # Re-index mask
         perm_mask = prev_k.unsqueeze(2).expand_as(self.mask)  # (batch_size, beam_size, num_nodes)
-        self.mask = self.mask.gather(1, perm_mask)
+        self.mask = self.mask.gather(1, perm_mask.long())
         # Mask newly added nodes
         self.update_mask(new_nodes)
 
@@ -117,6 +117,6 @@ class Beamsearch(object):
         k = k.type(torch.long).to(self.device)
         hyp = -1 * torch.ones(self.batch_size, self.num_nodes, dtype=torch.long).to(self.device)
         for j in range(len(self.prev_Ks) - 1, -2, -1):
-            hyp[:, j + 1] = self.next_nodes[j + 1].gather(1, k).view(1, self.batch_size)
+            hyp[:, j + 1] = self.next_nodes[j + 1].gather(1, k.long()).view(1, self.batch_size)
             k = self.prev_Ks[j].gather(1, k)
         return hyp
